refactor(apple-wizard): move check-installed-cli to separate file
(cherry picked from commit bcd268ef35b07503f2d0a0fae07f42a07295d5ce)
